+++ src/Term/FingerprintProvider.php
@@ -3,7 +3,7 @@
 namespace Wikibase\DataModel\Term;
 
 /**
- * @since 0.7.3
+ * @deprecated since 3.0.0, use EntityTermsProvider instead, will be dropped in 4.0.0
  *
  * @licence GNU GPL v2+
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
@@ -11,6 +11,8 @@ namespace Wikibase\DataModel\Term;
 interface FingerprintProvider {
 
 	/**
+	 * @deprecated since 3.0.0, use getEntityTerms instead, will be dropped in 4.0.0
+	 *
 	 * @return Fingerprint
 	 */
 	public function getFingerprint();