+++ RELEASE-NOTES.md
@@ -1,5 +1,10 @@
 # Wikibase DataModel release notes
 
+## Version 3.0.0 (dev)
+
+* Renamed `Fingerprint` to `EntityTerms` including all related methods, e.g. `Entity::getEntityTerms`
+* Kept `Fingerprint`, `FingerprintPatcher` and `FingerprintProvider` as deprecated aliases and/or subclasses
+
 ## Version 2.6.0 (dev)
 
 * Empty strings are now detected as invalid in the `SiteLink` constructor